# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface DeleteResult {
  success: boolean;
}

export interface AuthInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface User {
  id: string;
  email: string;
  createdAt: string;
}

export interface AuthResult {
  token: string;
  user: User;
}

export interface Skill {
  id: string;
  userId: string;
  name: string;
  description: string;
  instructions: string;
  /** @nullable */
  tool: string | null;
  enabled: boolean;
  priority: number;
  triggerExamples: string[];
  usageCount: number;
  /** @nullable */
  lastUsedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface SkillInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  instructions: string;
  tool?: string;
  enabled?: boolean;
  priority?: number;
  triggerExamples?: string[];
}

export interface SkillUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  instructions?: string;
  tool?: string;
  enabled?: boolean;
  priority?: number;
  triggerExamples?: string[];
}

export interface MatchInput {
  message: string;
}

export interface MatchResult {
  matched: boolean;
  confidence: number;
  skill?: Skill;
  reason?: string;
}

export interface GenerateInput {
  prompt: string;
}

export interface GeneratedSkill {
  name: string;
  description: string;
  instructions: string;
  triggerExamples: string[];
}

export interface TestInput {
  message: string;
}

export interface TestResult {
  wouldTrigger: boolean;
  triggerScore?: number;
  injectedPrompt: string;
  sampleResponse: string;
}

export interface SkillStat {
  id: string;
  name: string;
  usageCount: number;
  /** @nullable */
  lastUsedAt?: string | null;
}

export interface Stats {
  totalSkills: number;
  activeSkills: number;
  totalUses: number;
  topSkills: SkillStat[];
}