# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: iskills2-auth
    description: iSkills2 authentication
  - name: iskills2-skills
    description: iSkills2 skill management
  - name: iskills2-stats
    description: iSkills2 dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ── Auth ────────────────────────────────────────────────────────────────────

  /iskills2/auth/register:
    post:
      operationId: registerUser
      tags: [iskills2-auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /iskills2/auth/login:
    post:
      operationId: loginUser
      tags: [iskills2-auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /iskills2/auth/me:
    get:
      operationId: getMe
      tags: [iskills2-auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Skills ──────────────────────────────────────────────────────────────────

  /iskills2/skills:
    get:
      operationId: listSkills
      tags: [iskills2-skills]
      summary: List all skills for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Skills list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    post:
      operationId: createSkill
      tags: [iskills2-skills]
      summary: Create a skill
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /iskills2/skills/match:
    post:
      operationId: matchSkill
      tags: [iskills2-skills]
      summary: Find the best matching skill for a user message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchInput"
      responses:
        "200":
          description: Match result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /iskills2/skills/generate:
    post:
      operationId: generateSkill
      tags: [iskills2-skills]
      summary: AI-generate skill fields from a natural language description
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateInput"
      responses:
        "200":
          description: Generated skill fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedSkill"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /iskills2/skills/{id}:
    get:
      operationId: getSkill
      tags: [iskills2-skills]
      summary: Get a single skill
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    patch:
      operationId: updateSkill
      tags: [iskills2-skills]
      summary: Update a skill
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    delete:
      operationId: deleteSkill
      tags: [iskills2-skills]
      summary: Delete a skill
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /iskills2/skills/{id}/test:
    post:
      operationId: testSkill
      tags: [iskills2-skills]
      summary: Test a skill against a sample message and get a preview response
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestInput"
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /iskills2/skills/{id}/use:
    post:
      operationId: recordSkillUse
      tags: [iskills2-skills]
      summary: Record a skill usage event for analytics
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Stats ────────────────────────────────────────────────────────────────────

  /iskills2/stats:
    get:
      operationId: getStats
      tags: [iskills2-stats]
      summary: Get dashboard stats — total skills, active count, top skills by usage
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # ── Auth ──────────────────────────────────────────────────────────────────

    AuthInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time

    # ── Skills ────────────────────────────────────────────────────────────────

    Skill:
      type: object
      required: [id, userId, name, description, instructions, tool, enabled, priority, triggerExamples, usageCount, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        description:
          type: string
        instructions:
          type: string
        tool:
          type: ["string", "null"]
        enabled:
          type: boolean
        priority:
          type: integer
        triggerExamples:
          type: array
          items:
            type: string
        usageCount:
          type: integer
        lastUsedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    SkillInput:
      type: object
      required: [name, description, instructions]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        instructions:
          type: string
          minLength: 1
        tool:
          type: string
        enabled:
          type: boolean
        priority:
          type: integer
        triggerExamples:
          type: array
          items:
            type: string

    SkillUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        instructions:
          type: string
        tool:
          type: string
        enabled:
          type: boolean
        priority:
          type: integer
        triggerExamples:
          type: array
          items:
            type: string

    MatchInput:
      type: object
      required: [message]
      properties:
        message:
          type: string

    MatchResult:
      type: object
      required: [matched, confidence]
      properties:
        matched:
          type: boolean
        confidence:
          type: number
        skill:
          $ref: "#/components/schemas/Skill"
        reason:
          type: string

    GenerateInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string

    GeneratedSkill:
      type: object
      required: [name, description, instructions, triggerExamples]
      properties:
        name:
          type: string
        description:
          type: string
        instructions:
          type: string
        triggerExamples:
          type: array
          items:
            type: string

    TestInput:
      type: object
      required: [message]
      properties:
        message:
          type: string

    TestResult:
      type: object
      required: [wouldTrigger, injectedPrompt, sampleResponse]
      properties:
        wouldTrigger:
          type: boolean
        triggerScore:
          type: number
        injectedPrompt:
          type: string
        sampleResponse:
          type: string

    Stats:
      type: object
      required: [totalSkills, activeSkills, totalUses, topSkills]
      properties:
        totalSkills:
          type: integer
        activeSkills:
          type: integer
        totalUses:
          type: integer
        topSkills:
          type: array
          items:
            $ref: "#/components/schemas/SkillStat"

    SkillStat:
      type: object
      required: [id, name, usageCount]
      properties:
        id:
          type: string
        name:
          type: string
        usageCount:
          type: integer
        lastUsedAt:
          type: ["string", "null"]
          format: date-time