# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthResult {
  token: string;
  user: User;
}