# lib/api-zod/src/generated/types/generatedSkill.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedSkill {
  name: string;
  description: string;
  instructions: string;
  triggerExamples: string[];
}