# lib/api-zod/src/generated/types/matchResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Skill } from './skill';

export interface MatchResult {
  matched: boolean;
  confidence: number;
  skill?: Skill;
  reason?: string;
}