# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Skill {
  id: string;
  userId: string;
  name: string;
  description: string;
  instructions: string;
  /** @nullable */
  tool: string | null;
  enabled: boolean;
  priority: number;
  triggerExamples: string[];
  usageCount: number;
  /** @nullable */
  lastUsedAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}