# lib/api-zod/src/generated/types/skillInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SkillInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  instructions: string;
  tool?: string;
  enabled?: boolean;
  priority?: number;
  triggerExamples?: string[];
}