# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SkillStat } from './skillStat';

export interface Stats {
  totalSkills: number;
  activeSkills: number;
  totalUses: number;
  topSkills: SkillStat[];
}